refactor: reuse JWinPackager for compression
